[Fatfs] Implement fatfs synchronization object and simulate linux ramdisk for ut case

1. Add fatfs lseek implementation
2. remove ffsystem.c and implement sync object for re-entrant (fatfs.c)
3. simulate Linux ram-disk for ut (diskio.c)
4. add new ut case for fatfs

See merge request !1015

+++ test/testcase/framework/vfs_test.c
@@ -33,7 +33,7 @@ static int test_ioctl(file_t *node, int cmd, unsigned long arg)
     return -123;
 }
 
-static int test_lseek(file_t *fp, off_t off, int whence)
+static off_t test_lseek(file_t *fp, off_t off, int whence)
 {
     return -123;
 }
